WBS-14 Add css. Cleanup code
