+++ ConsoleApplication1.vcxproj
@@ -146,12 +146,17 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="src\LoggingPage.cpp" />
+    <ClCompile Include="src\HelloApplication.cpp" />
+    <ClCompile Include="src\WebBankingApplication.cpp" />
     <ClCompile Include="src\Main.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="include\HelloApplication.h" />
     <ClInclude Include="include\Hello.h" />
-    <ClInclude Include="include\LoggingPage.h" />
+    <ClInclude Include="include\WebBankingApplication.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="css\FrontPage.css" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
