+++ include/HelloApplication.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <Wt/WApplication.h>
+#include <Wt/WBreak.h>
+#include <Wt/WContainerWidget.h>
+#include <Wt/WLineEdit.h>
+#include <Wt/WPushButton.h>
+#include <Wt/WText.h>
+
+// This application is only for
+// educational purposes as a simple "Hello World Application"
+class HelloApplication : public Wt::WApplication
+{
+public:
+    HelloApplication(const Wt::WEnvironment& env);
+
+private:
+    Wt::WLineEdit* nameEdit_;
+    Wt::WText* greeting_;
+
+    void greet();
+};
