--- include/LoggingPage.h
@@ -1,10 +1,0 @@
-#pragma once
-#include <Wt/WContainerWidget.h>
-#include <C:/Users/QVM386/Desktop/web-banking-app/Wt 4.3.1 msvs2019 x64/include/Wt/WContainerWidget.h>
-
-class LoggingPage : public Wt::WContainerWidget
-{
-public:
-	LoggingPage();
-};
-
